Release 0.1-201511191609: Fix return type of java.lang.Iterable.iterator()
Apparently, scala compiles def A = new B() { ... } down to a method that
actually returns Object, not B. I have no idea how that would still
conform to the implemented interface, since it doesn't. This fix
gives the methods an explicit return type, so the poor Java compiler
can make sense of it.